相对稳定版本

+++ app/src/main/java/com/munin/mhsocket/socket/entity/SocketInput.java
@@ -133,6 +133,7 @@ public class SocketInput {
                 e1.printStackTrace();
             }
             close();
+            if(null!=client)
             client.destroy();
         }
     }
